feat: Add llama_dart talk() function
This commit adds a new function called `talk()` to the `llama_dart` library. The `talk()` function initializes the Llama context with default parameters and returns a string that indicates the initialization status. This change also includes updates to the main.dart file to use the `talk()` function instead of the previous sum functions.